Support the list of service providers to be `iterable`
This requires quite some rework, as iterable's might
be non traversable, that means we can iterate only once.

We therefore switch the current logic to iterate the
providers once, but that comes with a small cost:
We need to store the (original) factories in a lookup table,
to be able to support an extension defined by an early
service provider for a factory of a later service provider
(which needs to work as per the spec).